Generate a package report in HTML

Create a report in HTML

create_pkg_report(
  pkgname,
  from = NULL,
  to = NULL,
  path = paste0(pkgname, " - summary report", from, "_", to),
  gh = NULL
)

Arguments

pkgname
String containing the name of the package.

from
String in the format of YYYY-MM-DD specifying the start date of the reporting period. If set to NULL, the first day of the month of the system date will be used.

to
String in the format of YYYY-MM-DD specifying the end date of the reporting period. If set to NULL, the last day of the month of the system date will be used.

path
Pass the file path and the desired file name, excluding the file extension. For example, "my package report".

gh
String in the format of owner/repo to specify the GitHub repository that holds the package. Defaults to NULL by default, where GitHub statistics will be omitted.

Value

Opens and saves a static HTML file in the active working directory with the file name specified in path.