Get issues information from GitHub for the given time period.

Get information on issues from the specified GitHub repository. This is a wrapper around the 'gh' package. Original code taken from https://github.com/jennybc/analyze-github-stuff-with-r/.

get_gh_issues(owner, repo, start_date = NULL, end_date = NULL)

Arguments

owner
String specifying the owner of the repository.

repo
String specifying the name of the repo.

start_date
String specifying start date of reporting period to filter by, in the format YYYY-MM-DD. NULL by default, where no filters will be applied.

end_date
String specifying end date of reporting period to filter by, in the format YYYY-MM-DD. NULL by default, where no filters will be applied.

Examples

# \donttest{
get_gh_issues(owner = "martinctc", repo = "rwa")

#> No encoding supplied: defaulting to UTF-8.

#> # A tibble: 11 × 63
#>    url     repository_url  labels_url  comments_url  events_url  html_url     id
#>    <chr>   <chr>           <chr>       <chr>         <chr>       <chr>     <int>
#>  1 https:… https://api.gi… https://ap… https://api.… https://ap… https:/… 8.94e8
#>  2 https:… https://api.gi… https://ap… https://api.… https://ap… https:/… 8.64e8
#>  3 https:… https://api.gi… https://ap… https://api.… https://ap… https:/… 8.55e8
#>  4 https:… https://api.gi… https://ap… https://api.… https://ap… https:/… 8.53e8
#>  5 https:… https://api.gi… https://ap… https://api.… https://ap… https:/… 8.18e8
#>  6 https:… https://api.gi… https://ap… https://api.… https://ap… https:/… 8.02e8
#>  7 https:… https://api.gi… https://ap… https://api.… https://ap… https:/… 7.48e8
#>  8 https:… https://api.gi… https://ap… https://api.… https://ap… https:/… 7.37e8
#>  9 https:… https://api.gi… https://ap… https://api.… https://ap… https:/… 7.33e8
#> 10 https:… https://api.gi… https://ap… https://api.… https://ap… https:/… 7.14e8
#> 11 https:… https://api.gi… https://ap… https://api.… https://ap… https:/… 6.03e8
#> # … with 56 more variables: node_id <chr>, number <int>, title <chr>,
#> #   labels <list>, state <chr>, locked <lgl>, assignees <list>,
#> #   milestone <lgl>, comments <int>, created_at <date>, updated_at <date>,
#> #   closed_at <date>, author_association <chr>, active_lock_reason <lgl>,
#> #   body <chr>, performed_via_github_app <lgl>, user.login <chr>,
#> #   user.id <int>, user.node_id <chr>, user.avatar_url <chr>,
#> #   user.gravatar_id <chr>, user.url <chr>, user.html_url <chr>, …

# }